Fixed replace file algorithm
Previously the get proper filepath was calculated *BEFORE* the
optimization. Now this is fixed and the proper filepath is calculated
after the optimization.

Example: Consider a file `foo.webp` and we upload a file named
`foo.png`. Previously as the filepath was calculated before the
optimization, the upload would replace `foo.webp`. Now the proper
filepath will be `foo-1.webp`. Yay!